+++ demo/src/app/components/+layout/sidebar/sidebar.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-layout-sidebar',
+    templateUrl: './sidebar.component.html',
+    styleUrls: ['../layout-section.scss']
+})
+export class DemoLayoutSidebarComponent implements OnInit {
+    isolated = false;
+
+    width = '';
+
+    constructor() {}
+
+    ngOnInit(): void {}
+}