+++ src/nav/icon-nav/icon-nav-link.directive.ts
@@ -1,13 +1,21 @@
-import { Component, Directive, ElementRef, Renderer2, Input, HostBinding } from '@angular/core';
+import { Component, Directive, ElementRef, Renderer2, Input, HostBinding, ChangeDetectorRef } from '@angular/core';
 import { inputValueToBoolean } from '../../util/helpers';
 
-@Directive({
-    selector: '[thyIconNavLink]'
+@Component({
+    selector: '[thyIconNavLink]',
+    template: '<ng-content></ng-content><thy-icon *ngIf="icon" [thyIconName]="icon"></thy-icon>'
 })
-export class ThyIconNavLinkDirective {
+export class ThyIconNavLinkComponent {
     @HostBinding('class.active') navLinkActive = false;
 
     @HostBinding('class.thy-icon-nav-link') navLinkClass = true;
+
+    icon: string;
+
+    @Input()
+    set thyIconNavLinkIcon(icon: string) {
+        this.icon = icon;
+    }
 
     @Input()
     set thyIconNavLinkActive(active: string) {